Force client to fail in case of any problem found

+++ scripts/redeploy_agent.ps1
@@ -11,6 +11,8 @@ if ($pids -ne $null) {
 
 & taskkill.exe /F /IM eclipsec.exe 
 Start-Sleep -Seconds 2.5
-Remove-Item -Recurse -Force agent -ErrorAction Stop
+if (Test-Path agent) {
+	Remove-Item -Recurse -Force agent -ErrorAction Stop
+}
 Expand-Archive -Path "org.eclipse.rcptt.cloud.agent.product-win32.win32.x86_64.zip" -DestinationPath "agent"
 Start-Process -NoNewWindow agent\eclipsec.exe -ArgumentList "-consoleLog -serverHost 172.29.1.229  -serverPort 5009 -data agent\ws -vm `"C:\Users\Test\Documents\jdk-21.0.6.7-hotspot\bin\java.exe`" -vmargs -XX:+HeapDumpOnOutOfMemoryError -Xlog:gc"
